Add files via upload
This code creates a scale filter associated with the Daubechies wavelet with "n/2" vanish moments and permits generate visualizations to scale and wavelet function associated with created scale filter. The use is very easy: the unique input is the number "n" interpreted as the length the desirable filter or equivalently as twice the number of vanish moments.